mattgpt-react/backend/lambdas/requirements.txt
pinecone-client==3.0.0
openai==1.12.0
boto3==1.34.0
sentence-transformers==2.3.1
torch==2.1.2
